change timer steering
What we desire is long-term stability at the expense of short-term
accuracy.  Therefore, don't steer the timer based on local measurements
(which has a very low precision, counts per wwvb second is 1/1000
or 1000ppm), but instead base it on long-term measurements (e.g., counts
per wwvb day or so, 12ppb)

The resolution of the divisor is 1 clock per 32.678 seconds, or 6ppb when
the clock is nominally 5MHz.